Extract shared state and controller_service fixtures to conftest.py
Move duplicated ControllerTransitions fixture from test_service.py and
test_dashboard.py into conftest.py. Add MockController class and
controller_service fixture for tests needing a ControllerServiceImpl
with mock dependencies. Replace local _make_test_entrypoint with the
existing conftest make_test_entrypoint.
